V1/requirements.txt
pandas=2.3.3
numpy=2.3.3
gradio=5.49.1
python-dotenv=1.2.1
langchain-core=1.0.2
langchain-openai=1.0.1
openai=2.6.1
transformers=4.57.1